refactor(taxis): split behavior.rs into per-domain config modules
