+++ crates/taxis/src/config/behavior/mod.rs
@@ -1,0 +1,23 @@
+//! Deployment-tunable behavior configuration types.
+
+mod api;
+mod daemon;
+mod jwt;
+mod knowledge;
+mod messaging;
+mod nous;
+mod provider;
+mod timeouts;
+mod tools;
+mod tuning;
+
+pub use api::ApiLimitsConfig;
+pub use daemon::DaemonBehaviorConfig;
+pub use jwt::JwtSettings;
+pub use knowledge::KnowledgeConfig;
+pub use messaging::MessagingConfig;
+pub use nous::NousBehaviorConfig;
+pub use provider::{AnthropicConfig, PromptCacheMode, ProviderBehaviorConfig};
+pub use timeouts::{CapacityConfig, RetrySettings, TimeoutsConfig};
+pub use tools::ToolLimitsConfig;
+pub use tuning::TuningConfig;
